+++ ExpenseManagement/ExpenseManagement/Controllers/UserController.cs
@@ -1,5 +1,6 @@
 ﻿using ExpenseManagement.Models;
 using ExpenseManagement.Services.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
@@ -7,6 +8,7 @@ namespace ExpenseManagement.Controllers
 {
     [Route("api/users")]
     [ApiController]
+    [Authorize(AuthenticationSchemes = "Bearer", Roles = "admin")]
     public class UserController : ControllerBase
     {
         private readonly IUserService _userService;
@@ -20,12 +22,17 @@ namespace ExpenseManagement.Controllers
        
 
         [HttpPut("{userId}")]
-        public ActionResult UpdateExpense(UserToUpdateDto userToUpdateDto, string userId)
+        public ActionResult UpdateUser(UserToUpdateDto userToUpdateDto, string userId)
         {
 
 
             if (userToUpdateDto == null)
                return BadRequest();
+
+            var user = _userService.GetUser(userId);
+
+            if (user == null)
+                return NotFound();
 
             _userService.UpdateUser(userToUpdateDto, userId);
 
@@ -34,15 +41,46 @@ namespace ExpenseManagement.Controllers
         }
 
         [HttpDelete("{userId}")]
-        public ActionResult DeleteExpense(string userId)
+        public ActionResult DeleteUser(string userId)
         {
 
-            
+            var user = _userService.GetUser(userId);
+
+            if (user == null)
+                return NotFound();
+
 
             _userService.DeleteUser(userId);
 
 
             return Ok("User deleted successfully");
         }
+
+        [HttpGet("{userId}")]
+        public ActionResult GetUser(string userId)
+        {
+
+            var user = _userService.GetUser(userId);
+
+            if (user == null)
+                return NotFound();
+
+
+            return Ok(user);
+        }
+
+        [HttpGet()]
+        public ActionResult GetUsers()
+        {
+
+            var users = _userService.GetUsers();
+
+            if (users == null || users.Count() == 0)
+                return NotFound();
+
+
+            return Ok(users);
+        }
+
     }
 }
